Split out utils module, move files from web.
