+++ deepwell/src/methods/mod.rs
@@ -35,7 +35,8 @@ mod prelude {
         RenderService, RequestFetchService, RevisionService, ScoreService,
         ServiceContext, SiteService, TextService, UserService, VoteService,
     };
-    pub use crate::web::{utils::error_response, CuidReference, HttpUnwrap, Reference};
+    pub use crate::utils::error_response;
+    pub use crate::web::{CuidReference, HttpUnwrap, Reference};
     pub use chrono::prelude::*;
     pub use sea_orm::{ConnectionTrait, TransactionTrait};
     pub use std::convert::TryFrom;
