+++ deepwell/src/utils/category.rs
@@ -1,5 +1,5 @@
 /*
- * web/category.rs
+ * utils/category.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2022 Wikijump Team
