+++ deepwell/src/utils/user.rs
@@ -1,5 +1,5 @@
 /*
- * util.rs
+ * utils/user.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2022 Wikijump Team
@@ -18,40 +18,12 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-use crate::services::{Error, Result};
-use chrono::{DateTime, FixedOffset, Utc};
-use std::future::Future;
+use crate::utils::replace_in_place;
 use wikidot_normalize::normalize;
-
-pub type DateTimeWithTimeZone = DateTime<FixedOffset>;
-
-pub fn replace_in_place(string: &mut String, pattern: &str, replacement: &str) {
-    while let Some(index) = string.find(pattern) {
-        let end = index + replacement.len();
-
-        string.replace_range(index..end, replacement);
-    }
-}
-
-lazy_static! {
-    pub static ref UTC: FixedOffset = FixedOffset::east(0);
-}
-
-#[inline]
-pub fn now() -> DateTimeWithTimeZone {
-    Utc::now().with_timezone(&*UTC)
-}
 
 pub fn get_user_slug<S: Into<String>>(name: S) -> String {
     let mut slug = name.into();
     replace_in_place(&mut slug, ":", "-");
     normalize(&mut slug);
     slug
-}
-
-pub async fn find_or_error<F, T>(future: F) -> Result<T>
-where
-    F: Future<Output = Result<Option<T>>>,
-{
-    future.await?.ok_or(Error::NotFound)
 }
